+++ actionview/test/actionpack/controller/render_test.rb
@@ -605,6 +605,25 @@ class TestController < ActionController::Base
     render action: "calling_partial_with_layout", layout: "layouts/partial_with_layout"
   end
 
+  def render_renderable
+    locals = params[:locals]
+    format = params[:format]
+
+    renderable = Object.new
+    renderable.singleton_class.define_method :render_in do |view_context, **, &|
+      view_context.render inline: <<~ERB.strip, locals: locals
+        Hello, <%= name %>
+      ERB
+    end
+    if format.present?
+      renderable.singleton_class.define_method :format do
+        format.to_sym
+      end
+    end
+
+    render renderable: renderable
+  end
+
   before_action only: :render_with_filters do
     request.format = :xml
   end
@@ -713,6 +732,7 @@ class RenderTest < ActionController::TestCase
     get :render_and_redirect, to: "test#render_and_redirect"
     get :render_call_to_partial_with_layout, to: "test#render_call_to_partial_with_layout"
     get :render_call_to_partial_with_layout_in_main_layout_and_within_content_for_layout, to: "test#render_call_to_partial_with_layout_in_main_layout_and_within_content_for_layout"
+    get :render_renderable, to: "test#render_renderable"
     get :render_custom_code, to: "test#render_custom_code"
     get :render_file_from_template, to: "test#render_file_from_template"
     get :render_file_not_using_full_path, to: "test#render_file_not_using_full_path"
@@ -1497,6 +1517,13 @@ class RenderTest < ActionController::TestCase
   def test_render_call_to_partial_with_layout_in_main_layout_and_within_content_for_layout
     get :render_call_to_partial_with_layout_in_main_layout_and_within_content_for_layout
     assert_equal "Before (Anthony)\nInside from partial (Anthony)\nAfter\nBefore (David)\nInside from partial (David)\nAfter\nBefore (Ramm)\nInside from partial (Ramm)\nAfter", @response.body
+  end
+
+  def test_render_renderable
+    get :render_renderable, params: { format: :html, locals: { name: "Local" } }
+
+    assert_equal "Hello, Local", @response.body
+    assert_equal "text/html", @response.media_type
   end
 
   def with_annotations_enabled